requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
billiard==3.6.3.0
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-filter==2.2.0
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
dnspython==1.16.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
